fix(secondgand market): apply mentor feedback
